+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- [Build] Drop webui-download.cmake npm-resolver patch (fixed upstream in #23064)
+
 
 ## [1.33.0] - 2026-05-15